requirements.txt
pandas>=2.0
python-dateutil>=2.8
tqdm>=4.66
selenium>=4.20
PySide6>=6.7
pillow>=10.0
python-dotenv>=1.0
pytest>=8.0
